backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
aws==0.2.5
bcrypt==5.0.0
boto==2.49.0
boto3==1.40.70
botocore==1.40.70
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cryptography==46.0.3
cycler==0.12.1
decorator==5.2.1
Deprecated==1.3.1
dnspython==2.7.0
exceptiongroup==1.3.0
fabric==3.2.2
fastapi==0.121.1
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.10.0
h11==0.16.0
idna==3.11
importlib_resources==6.5.2
invoke==2.2.1
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.12.0.88
packaging==25.0
paramiko==4.0.0
pika==1.3.2
pillow==11.3.0
polars==1.35.2
polars-runtime-32==1.35.2
prettytable==3.16.0
psutil==7.1.3
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pymongo==4.15.3
PyNaCl==1.6.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
s3transfer==0.14.0
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.227
ultralytics-thop==2.0.18
urllib3==1.26.20
uvicorn==0.38.0
wcwidth==0.2.14
wrapt==2.0.1
yt-dlp==2025.10.14
zipp==3.23.0
